+++ next-docs/public/examples/carousel/Rtl.tsx
@@ -2,10 +2,10 @@ import { Carousel } from '@heathmont/moon-components';
 import { rem } from '@heathmont/moon-utils';
 import styled from 'styled-components';
 
-const ExampleContent = styled.div(({ theme: { color, radius } }) => ({
-  width: '20rem',
-  height: '12rem',
-  borderRadius: rem(radius.small),
+const ExampleContent = styled.div(({ theme: { color, newTokens } }) => ({
+  width: rem(320),
+  height: rem(192),
+  borderRadius: newTokens.borderRadius.xsmall,
   background: color.goku[80],
   display: 'flex',
   alignItems: 'center',
@@ -13,12 +13,11 @@ const ExampleContent = styled.div(({ theme: { color, radius } }) => ({
 }));
 
 const Container = styled.div({
-  maxWidth: '600px',
+  maxWidth: rem(600),
 });
 
 const Example = () => {
   const items = Array.from({ length: 25 }, (_, index) => index);
-
   return (
     <div className="flex justify-around items-center w-full">
       <Container>